Make XCUIElement available to both L10n and XCUITest
