+++ firefox-ios/firefox-ios-tests/Tests/L10nSnapshotTests/L10nBaseSnapshotTests.swift
@@ -131,57 +131,3 @@ class L10nBaseSnapshotTests: XCTestCase {
         mozWaitForElementToNotExist(progressIndicator, timeout: 30.0)
     }
 }
-
-extension XCUIElement {
-    func tapOnApp() {
-        coordinate(withNormalizedOffset: CGVector(dx: 0.5, dy: 0.5)).tap()
-    }
-
-    func mozWaitElementHittable(timeout: Double) {
-        let predicate = NSPredicate(format: "exists == true && hittable == true")
-        let expectation = XCTNSPredicateExpectation(predicate: predicate, object: self)
-        let result = XCTWaiter().wait(for: [expectation], timeout: timeout)
-        XCTAssertEqual(result, .completed, "Element did not become hittable in time.")
-    }
-
-    func mozWaitForElementToExist(timeout: TimeInterval? = TIMEOUT) {
-        let startTime = Date()
-        guard exists else {
-            while !exists {
-                if let timeout = timeout, Date().timeIntervalSince(startTime) > timeout {
-                    XCTFail("Timed out waiting for element \(self) to exist in \(timeout) seconds")
-                    break
-                }
-                usleep(10000)
-            }
-            return
-        }
-    }
-
-    /// Waits for the UI element and then taps if it exists.
-    func waitAndTap(timeout: TimeInterval? = TIMEOUT) {
-        self.mozWaitForElementToExist(timeout: timeout)
-        self.tap()
-    }
-
-    /// Waits for the UI element and then taps and types the provided text if it exists.
-    func tapAndTypeText(_ text: String, timeout: TimeInterval? = TIMEOUT) {
-        self.mozWaitForElementToExist(timeout: timeout)
-        self.tap()
-        self.typeText(text)
-    }
-
-    func pressWithRetry(duration: TimeInterval, timeout: TimeInterval = TIMEOUT, element: XCUIElement) {
-        self.mozWaitForElementToExist(timeout: timeout)
-        self.press(forDuration: duration)
-        var attempts = 5
-        while !element.exists && attempts > 0 {
-            self.press(forDuration: duration)
-            attempts -= 1
-        }
-
-        if !element.exists {
-            XCTFail("\(element) is not visible after \(attempts) attempts")
-        }
-    }
-}
